pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.19.0",
    "resend>=0.8.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
    "pypdf>=4.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "requests>=2.31.0",
    "anthropic>=0.18.0",
]
